rewrite with ts post feature and group

+++ app/src/components/Group.container.tsx
@@ -1,16 +1,17 @@
 import React, {useEffect, useState} from 'react';
-import {useNavigate, useLocation, useParams} from 'react-router-dom';
+import {useNavigate, useParams} from 'react-router-dom';
 import axios from 'axios';
 import {useToken} from '../utils/useToken';
 import { GroupDashboardComponent } from './GroupDashboard.component';
+import {GroupType, UserType} from '../types';
 
 export const GroupContainer = () => {
   const {token} = useToken();
   const {name} = useParams();
   const navigate = useNavigate();
-  const [users, setUsers] = useState();
-  const [currentUser, setCurrentUser] = useState([]);
-  const [group, setGroup] = useState();
+  const [users, setUsers] = useState<UserType[]>([]);
+  const [currentUser, setCurrentUser] = useState<UserType>();
+  const [group, setGroup] = useState<GroupType>();
   const [error, setError] = useState('');
   const [loading, setLoading] = useState(false);
 
@@ -23,7 +24,7 @@ export const GroupContainer = () => {
       setError('Sorry, can\'t load an user data');
     }   
   };
-  const onLoadGroupUsers = async (id) => {
+  const onLoadGroupUsers = async (id: string) => {
     try {
       const resp =  await axios.get(`${process.env.PUBLIC_URL}/api/user_group/group/${id}`, 
       { headers: {"Authorization" : `Bearer ${token}`}});
@@ -41,7 +42,7 @@ export const GroupContainer = () => {
     }
     setLoading(false);
   };
-  const onLoadUsers = async (userIds) => {
+  const onLoadUsers = async (userIds: string[]) => {
     setLoading(true);
     try {
       const resp =  await axios.get(`${process.env.PUBLIC_URL}/api/users/group/${userIds}`);
@@ -53,6 +54,7 @@ export const GroupContainer = () => {
     setLoading(false);
   };
   const onJoinGroup = async () => {
+    if(!group) return;
     setLoading(true);
     try {
       await axios.post(`${process.env.PUBLIC_URL}/api/user_group/${token}/create`, {groupId: group._id},
@@ -66,6 +68,7 @@ export const GroupContainer = () => {
     setLoading(false);
   };
   const onLeaveGroup = async () => {
+    if(!group) return;
     setLoading(true);
     try {
       await axios.delete(`${process.env.PUBLIC_URL}/api/user_group/${token}/delete/${group._id}`,
@@ -86,12 +89,12 @@ export const GroupContainer = () => {
 
   return (
     <GroupDashboardComponent group={group}
-                    users={users}
-                    onJoinGroup={onJoinGroup}
-                    onLeaveGroup={onLeaveGroup}
-                    currentUser={currentUser}
-                    error={error}
-                    loading={loading}/>
+                             users={users}
+                             onJoinGroup={onJoinGroup}
+                             onLeaveGroup={onLeaveGroup}
+                             currentUser={currentUser}
+                             error={error}
+                             loading={loading}/>
   );
 }
